Refactor: comment 주석 최신화

+++ src/main/java/company/board_project/domain/comment/service/CommentService.java
@@ -18,7 +18,6 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
-
 @Service
 @RequiredArgsConstructor
 @Transactional
@@ -28,13 +27,14 @@ public class CommentService {
     private final UserService userService;
     private final ContentService contentService;
 
-    // 댓글 등록
+    /*
+     * 댓글 등록
+     */
     public Comment createComment(
             Comment comment,
             Long contentId,
             Long userId) {
 
-        // 이미 등록된 이메일인지 확인
         Content content = contentService.findContent(contentId);
         User user = userService.findUser(userId);
 
@@ -44,7 +44,9 @@ public class CommentService {
         return commentRepository.save(comment);
     }
 
-    // 댓글 수정
+    /*
+     * 댓글 수정
+     */
     public Comment updateComment(
             Comment comment,
             Long commentId) {
@@ -52,9 +54,9 @@ public class CommentService {
 
         Comment findComment = findVerifiedComment(commentId); //ID로 멤버 존재 확인하고 comment 정보 반환
 
-//        User writer = userService.findUser(findComment.getUser().getUserId()); // 작성자 찾기
-/*        if (memberService.getLoginUser().getUserId() != writer.getUserId()) // 작성자와 로그인한 사람이 다를 경우
-            throw new BusinessLogicException(Exceptions.UNAUTHORIZED);*/
+        User writer = userService.findUser(findComment.getUser().getUserId()); // 작성자 찾기
+        if (userService.getLoginUser().getUserId() != writer.getUserId()) // 작성자와 로그인한 사람이 다를 경우
+            throw new BusinessLogicException(Exceptions.UNAUTHORIZED);
 
         Optional.ofNullable(comment.getComment())
                 .ifPresent(findComment::setComment);
@@ -62,35 +64,44 @@ public class CommentService {
         return commentRepository.save(findComment);
     }
 
-    // 댓글 단건 조회
+    /*
+     * 댓글 단건 조회
+     */
     public Comment findComment(long commentId) {
         return findVerifiedComment(commentId);
     }
 
-
-    // 댓글 전체 조회
+    /*
+     * 댓글 전체 조회
+     */
     public Page<Comment> findComments(int page, int size) {
         return commentRepository.findAll(PageRequest.of(page, size,
                 Sort.by("commentId").descending()));
     }
 
-    // 게시글 Id 단위 댓글 조회
+    /*
+     * 게시글 Id 단위 댓글 조회
+     */
     public List<Comment> findContentComments(int contentId) {
         return findVerifiedContentComments(contentId);
     }
 
-    // 댓글 삭제
+    /*
+     * 댓글 삭제
+     */
     public void deleteComment(long commentId) {
         Comment findComment = findVerifiedComment(commentId);
 
-//        User writer = userService.findUser(findComment.getUser().getUserId()); // 작성자 찾기
-      /*  if (memberService.getLoginMember().getMemberId() != writer.getMemberId()) // 작성자와 로그인한 사람이 다를 경우
-            throw new BusinessLogicException(Exceptions.UNAUTHORIZED);*/
+        User writer = userService.findUser(findComment.getUser().getUserId()); // 작성자 찾기
+        if (userService.getLoginUser().getUserId() != writer.getUserId()) // 작성자와 로그인한 사람이 다를 경우
+            throw new BusinessLogicException(Exceptions.UNAUTHORIZED);
 
         commentRepository.delete(findComment);
     }
 
-    // 댓글 존재 유무 확인
+    /*
+     * 댓글 존재 유무 확인
+     */
     public Comment findVerifiedComment(long commentId) {
         Optional<Comment> optionalComment = commentRepository.findById(commentId);
         Comment findComment =
@@ -99,7 +110,9 @@ public class CommentService {
         return findComment;
     }
 
-    // 게시글 존재 유무 확인
+    /*
+     * 게시글 존재 유무 확인
+     */
     public List<Comment> findVerifiedContentComments(long contentId) {
         List<Comment> findContentComments = commentRepository.findAllByContentId(contentId);
